fix: `Value.Assign` should work with `int32` and int64.
This applies to:

* `Value.Assign(int32)`
* `Value.Assign(uint32)`

* `Value.Assign(int64)`
* `Value.Assign(uint32)`

* `Value.Assign(int)`
* `Value.Assign(uint)`

refs #286

+++ value.go
@@ -83,7 +83,7 @@ func (v *Value) Assign(val interface{}) {
 		if val.(bool) {
 			i = 1
 		}
-		v.SetInt(i)
+		v.SetInt(int32(i))
 	case float32:
 		v.Assign(float64(val.(float32)))
 	case int:
